タイマーを管理するクラスを作成
